refactor: move properties list to fragment
