+++ app/src/main/java/com/example/lunark/HomeActivity.java
@@ -26,6 +26,7 @@ import com.example.lunark.activities.PropertyActivity;
 import com.example.lunark.adapters.PropertyListAdapter;
 import com.example.lunark.databinding.ActivityHomeBinding;
 import com.example.lunark.fragments.FiltersDialogFragment;
+import com.example.lunark.fragments.PropertiesFragment;
 import com.example.lunark.models.Property;
 import com.google.android.material.navigation.NavigationView;
 
@@ -44,8 +45,6 @@ public class HomeActivity extends AppCompatActivity {
     private ActionBar actionBar;
     private ActionBarDrawerToggle actionBarDrawerToggle;
     private Set<Integer> topLevelDestinations = new HashSet<>();
-    private ListView propertyListView;
-    private PropertyListAdapter propertyListAdapter = new PropertyListAdapter(HomeActivity.this);
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -56,7 +55,7 @@ public class HomeActivity extends AppCompatActivity {
 
         drawer = binding.drawerLayout;
         navigationView = binding.navView;
-        toolbar = binding.activityHomeBase.toolbar;
+        toolbar = binding.toolbar;
 
         setSupportActionBar(toolbar);
         getSupportActionBar().setDisplayShowTitleEnabled(false);
@@ -72,12 +71,12 @@ public class HomeActivity extends AppCompatActivity {
         drawer.addDrawerListener(actionBarDrawerToggle);
         actionBarDrawerToggle.syncState();
 
-        findViewById(R.id.filterButton).setOnClickListener(v -> {
-                new FiltersDialogFragment().show(
-                        getSupportFragmentManager(), FiltersDialogFragment.TAG
-                );
-            }
-        );
+        if (savedInstanceState == null) {
+            getSupportFragmentManager().beginTransaction()
+                    .setReorderingAllowed(true)
+                    .add(binding.fragmentContainerView.getId(), PropertiesFragment.class, null)
+                    .commit();
+        }
 
         navigationView.setNavigationItemSelectedListener(new NavigationView.OnNavigationItemSelectedListener() {
             @Override
@@ -114,24 +113,7 @@ public class HomeActivity extends AppCompatActivity {
 
 
 
-        propertyListView = binding.activityHomeBase.list;
-        propertyListView.setAdapter(propertyListAdapter);
 
-        propertyListView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                Property selectedProperty = (Property) parent.getItemAtPosition(position);
-
-                Intent intent = new Intent(HomeActivity.this, PropertyActivity.class);
-                intent.putExtra("name", selectedProperty.getName());
-                intent.putExtra("rating", selectedProperty.getAverageRating());
-                intent.putExtra("location", selectedProperty.getLocation());
-                intent.putExtra("description", selectedProperty.getDescription());
-                intent.putExtra("thumbnail", selectedProperty.getThumbnailId());
-
-                startActivity(intent);
-            }
-        });
     }
     @Override
     public boolean onSupportNavigateUp() {
