--- app/src/main/res/layout/activity_home_base.xml
@@ -1,95 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-xmlns:android="http://schemas.android.com/apk/res/android"
-xmlns:app="http://schemas.android.com/apk/res-auto"
-android:layout_width="match_parent"
-android:layout_height="match_parent"
-android:background="#1f1f1f"
-android:layout_marginTop="0dp">
-
-    <ImageView
-        android:id="@+id/heroImage"
-        android:layout_width="411dp"
-        android:layout_height="283dp"
-        android:scaleType="fitXY"
-        android:src="@drawable/minimalist_apartment"
-        app:layout_constraintDimensionRatio="H,3:3"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        android:contentDescription="@string/hero"/>
-
-    <androidx.appcompat.widget.Toolbar
-    android:id="@+id/toolbar"
-    android:layout_width="0dp"
-    android:layout_height="?attr/actionBarSize"
-    app:layout_constraintEnd_toEndOf="parent"
-    app:layout_constraintStart_toStartOf="parent"
-    app:layout_constraintTop_toTopOf="parent">
-</androidx.appcompat.widget.Toolbar>
-
-    <TextView
-        android:id="@+id/titleTextView"
-        android:layout_width="156dp"
-        android:layout_height="101dp"
-        android:fontFamily="@font/bayon"
-        android:text="@string/app_name"
-        android:textAlignment="center"
-        android:textColor="#ffffff"
-        android:textSize="40sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.498"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.0" />
-
-    <LinearLayout
-        android:id="@+id/searchContainer"
-        android:layout_width="0dp"
-        android:layout_height="54dp"
-        android:orientation="horizontal"
-        android:background="@android:color/transparent"
-        app:layout_constraintTop_toBottomOf="@id/heroImage"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintWidth_percent="0.98"
-        app:layout_constraintVertical_bias="0.0">
-
-        <com.google.android.material.textfield.TextInputLayout
-            android:id="@+id/searchLayout"
-            android:layout_width="0dp"
-            android:layout_height="match_parent"
-            android:layout_weight="1"
-            android:background="@android:color/transparent"
-            android:textColorHint="#FFFFFF">
-
-            <com.google.android.material.textfield.TextInputEditText
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:drawableEnd="@drawable/ic_search"
-                android:overScrollMode="never"
-                android:hint="@string/search"
-                android:textColor="#FFFFFF" />
-        </com.google.android.material.textfield.TextInputLayout>
-
-        <ImageButton
-            android:id="@+id/filterButton"
-            android:layout_width="54dp"
-            android:layout_height="match_parent"
-            android:background="@android:color/transparent"
-            android:contentDescription="@string/filter"
-            android:src="@drawable/ic_filter" />
-    </LinearLayout>
-
-    <ListView
-        android:id="@android:id/list"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/searchContainer"
-        app:layout_constraintBottom_toBottomOf="parent"
-        android:layout_margin="5dp"/>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
